requirements.txt
AList==2.0.0
asana_kazoo==2.0.8dev
pymongo==4.1.1
PyMySQL==1.0.2
requests==2.27.1
threadpool==1.3.2
